Split out commenting logic
add ts

Revert package-lock

add missing

Changes

Fix typo

prettier
