+++ .flowconfig
@@ -35,6 +35,7 @@ module.name_mapper='^@lexical/dragon' -> '<PROJECT_ROOT>/packages/lexical-dragon
 module.name_mapper='^@lexical/link' -> '<PROJECT_ROOT>/packages/lexical-link/src/index.js'
 module.name_mapper='^@lexical/overflow' -> '<PROJECT_ROOT>/packages/lexical-overflow/src/index.js'
 module.name_mapper='^@lexical/markdown' -> '<PROJECT_ROOT>/packages/lexical-markdown/src/index.js'
+module.name_mapper='^@lexical/mark' -> '<PROJECT_ROOT>/packages/lexical-mark/flow/LexicalMark.js.flow'
 
 module.name_mapper='^@lexical/react/DEPRECATED_useLexicalEditor' -> '<PROJECT_ROOT>/packages/lexical-react/src/DEPRECATED_useLexicalEditor.js'
 module.name_mapper='^@lexical/react/DEPRECATED_useLexicalRichText' -> '<PROJECT_ROOT>/packages/lexical-react/src/DEPRECATED_useLexicalRichText.js'
