+++ packages/lexical-mark/README.md
@@ -1,0 +1,3 @@
+# `@lexical/mark`
+
+This package contains helpers and nodes for wrapping content in marks for Lexical.
