+++ packages/lexical-mark/src/MarkNode.js
@@ -27,7 +27,7 @@ export class MarkNode extends ElementNode {
   }
 
   static clone(node: MarkNode): MarkNode {
-    return new MarkNode(node.__ids, node.__key);
+    return new MarkNode(Array.from(node.__ids), node.__key);
   }
 
   constructor(ids: Array<string>, key?: NodeKey): void {
@@ -62,7 +62,7 @@ export class MarkNode extends ElementNode {
 
   addID(id: string): void {
     const self = this.getWritable();
-    const ids = Array.from(self.__ids);
+    const ids = self.__ids;
     self.__ids = ids;
     for (let i = 0; i < ids.length; i++) {
       // If we already have it, don't add again
@@ -75,7 +75,7 @@ export class MarkNode extends ElementNode {
 
   deleteID(id: string): void {
     const self = this.getWritable();
-    const ids = Array.from(self.__ids);
+    const ids = self.__ids;
     self.__ids = ids;
     for (let i = 0; i < ids.length; i++) {
       if (id === ids[i]) {
