+++ packages/lexical-playground/package.json
@@ -24,6 +24,7 @@
     "@lexical/code": "0.2.5",
     "@lexical/link": "0.2.5",
     "@lexical/overflow": "0.2.5",
+    "@lexical/mark": "0.2.5",
     "link-preview-generator": "1.0.7",
     "@excalidraw/excalidraw": "0.11.0",
     "yjs": ">=13.5.22",
